fix: real compile errors uncovered by local mix compile
Three actual bugs that made every job fail when --warnings-as-errors
was disabled wouldn't have masked them — they're genuine compile errors,
not warnings:

  1. lib/tau/session/events.ex:3 — moduledoc heredoc contained
     `"session:#{id}"` which Elixir interpolates as a variable
     reference. There's no `id` in scope at module compile, so:
       error: undefined variable "id"
     Replaced with `"session:<id>"` literal placeholder.

  2. lib/tau/tools/builtin/read.ex:115 — `min/3` was used. Kernel only
     defines `min/2`. Switched to `Enum.min([limit, @max_lines, total])`.

  3. lib/tau/message.ex — three structs (Tau.Message.User,
     Tau.Message.Assistant, Tau.Message.ToolResult) plus the union
     module Tau.Message in one file. Tau.Message was DEFINED FIRST
     and pattern-matched on the structs that hadn't been defined yet
     (compile order is top-to-bottom). Moved the structs above the
     union module so Tau.Message can resolve them.

Plus a refactor of Tau.TUI to make the optional-Ratatouille pattern
actually work:

  4. lib/tau/tui.ex previously had `if Code.ensure_loaded?(Ratatouille.Runtime)
     do ... import Ratatouille.View ... end` directly in the module
     body. The if-condition was being evaluated correctly but the
     import was still triggering Elixir to look up Ratatouille.View
     during compile, even when the if-branch was dead. Split it:
     - lib/tau/tui.ex: thin shim that delegates to Tau.TUI.App.run/0
       at runtime, only after Code.ensure_loaded? returns true.
     - lib/tau/tui/app.ex: the entire Ratatouille.App implementation,
       wrapped in a top-level `if Code.ensure_loaded?(...)` so the
       module isn't even defined when ratatouille is missing.
     This pattern compiles cleanly without ratatouille and works at
     runtime when it's there.

Local mix compile (Elixir 1.17.3 / OTP 25, no deps fetched) now
proceeds past these errors and only fails on Finch.Response /
Phoenix.PubSub references that need actual deps. CI has the deps so
it should now actually run tests instead of dying at compile.
